Added fastq_screen module
